Se crteró el mpétogo eliminarCliente (sin validar que su sado sea 0) dentord el ServicioCRUDCliente

+++ src/main/java/com/acm/tiendaerick/paqueteClientes/dtoCliente/PagoRealizadoDTO.java
@@ -3,12 +3,14 @@ package com.acm.tiendaerick.paqueteClientes.dtoCliente;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 public record PagoRealizadoDTO(
         @PositiveOrZero(message = "El id del cliente no puede ser negativo")
         long id_cliente,
 
         @PositiveOrZero(message = "El saldo actual no puede ser negativo")
-        long saldo_actual,
+        BigDecimal saldo_actual,
 
         @NotBlank(message = "El mensaje de confirmación no puede estar vacío")
         String mensaje
